Create Detect cycle in a directed graph(DFS)
vis array will check whether all the nodes are visited or not , while pathVis array will check whether the node is visited in the current path. If it is visited in the current path then it is considered that there is cycle.

So the intuition is to reach a previously visited node again on the same path.

Note: If a directed graph contains a cycle, the node has to be visited again on the same path and not through different paths.


Time Complexity: O(V+E)+O(V) , where V = no. of nodes and E = no. of edges. There can be at most V components. So, another O(V) time complexity.

Space Complexity: O(2N) + O(N) ~ O(2N): O(2N) for two visited arrays and O(N) for recursive stack space.