Bipartite Graph
DFS Approach :- 

Here the concept is similar to that of detecting the cycle in undirected graph (i.e., traversing the entire graph and if we get to the same node then it implies that there is a cycle).

But here when you get to the visited node then you have to check that present node and the next connected node having same color or not. If same color is there then it is not a bipartite graph.

Time Complexity: O(V + 2E), Where V = Vertices, 2E is for total degrees as we traverse all adjacent nodes.

Space Complexity: O(3V) ~ O(V), Space for DFS stack space, colour array and an adjacency list.


*** BFS Approach ***
The concept is same as the detecting the cycle in  undirected graph(BFS) , here if we get to the visited node then we have to check the color of the present node and the next connected node. 

Time Complexity: O(V + 2E), Where V = Vertices, 2E is for total degrees as we traverse all adjacent nodes.

Space Complexity: O(3V) ~ O(V), Space for queue data structure, colour array and an adjacency list.