https://www.geeksforgeeks.org/problems/topological-sort/1
Algorithm Concept:
-> Use an in-degree array to track the number of incoming edges for each vertex.
-> Begin with vertices that have in-degree 0 (no incoming edges).
-> Use a queue (FIFO) to manage the process, iteratively removing vertices with in-degree 0 and updating the in-degree of their neighbors.
-> Add vertices with updated in-degree 0 to the queue, repeating until all vertices are processed.

Note: Points to remember when a node is popped out, indegrees for all its adjacent nodes are decreased by one and if any of them becomes 0, we push that node into the queue. Meanwhile, we include the current node in the answer immediately after it is popped out of the queue.


In simple:-
Steps include:
1. Compute in-degrees of vertices.
2. Initialize a queue with vertices having in-degree 0.
3. Process the queue, updating in-degrees of adjacent vertices.
4. Return the resulting topological order.

Time and Space Complexity:

Time complexity: O(V + E), where V is the number of vertices, and E is the number of edges.
Space complexity: O(V) for storing in-degrees and the queue.