https://www.geeksforgeeks.org/problems/detect-cycle-in-a-directed-graph/1
The approach is like the Kahn's algorithm. If the topological sort fails to process all nodes, it indicates the presence of a cycle. So, by this we can detect the cycle.


Comparison with DFS:
DFS relies on tracking visited nodes and path-visited nodes for backtracking.
BFS cannot use backtracking, necessitating a different approach.



Time Complexity: 
𝑂(𝑉+𝐸)
O(V+E) (similar to BFS for graphs).
Space Complexity: 
𝑂(𝑉)
O(V), accounting for the queue and in-degree array.