https://www.geeksforgeeks.org/problems/minimal-cost/0
Here, the problem is similar to frog jump but here instead of jumping 1 or 2 steps , we are jumping k steps.

If we generalize, we are making K calls, therefore, we can set a for loop to run from 1 to K, and in each iteration, we can make a function call, corresponding to a step. We will return the minimum step call after the loop.